Use decorator and context manager to mock strip api
