+++ corehq/apps/accounting/tests/utils.py
@@ -7,6 +7,14 @@ from corehq.apps.accounting.models import (
     SubscriptionAdjustment,
 )
 from corehq.apps.accounting.tests import generator
+import stripe
+from contextlib import contextmanager
+from unittest import mock
+from corehq.apps.accounting.tests.generator import (
+    FakeStripeCardManager,
+    FakeStripeCustomerManager,
+)
+from functools import wraps
 
 
 class DomainSubscriptionMixin(object):
@@ -54,3 +62,31 @@ class DomainSubscriptionMixin(object):
             cls.__accounts[domain].delete()
         finally:
             Subscription._get_active_subscription_by_domain.clear(Subscription, domain)
+
+
+@contextmanager
+def mocked_stripe_api_context(*args, **kwargs):
+    with mock.patch.object(stripe.Customer, 'modify_source') as mock_modify_source, \
+         mock.patch.object(stripe.Customer, 'list_sources') as mock_list_sources:
+        setup_stripe_common_mocks(mock_list_sources, mock_modify_source)
+        yield
+
+
+def mocked_stripe_api_decorator(func):
+    @wraps(func)
+    def wrapper(*args, **kwargs):
+        with mocked_stripe_api_context():
+            return func(*args, **kwargs)
+    return wrapper
+
+
+def setup_stripe_common_mocks(mock_list_sources, mock_modify_source):
+    def modify_source_side_effect(*args, **kwargs):
+        updated_metadata = {k: str(v) for k, v in kwargs['metadata'].items()}
+        FakeStripeCardManager.get_card_by_id(kwargs['id']).metadata.update(updated_metadata)
+
+    def list_sources_side_effect(*args, **kwargs):
+        return FakeStripeCustomerManager.get_customer_by_id(kwargs['customer']).cards
+
+    mock_modify_source.side_effect = modify_source_side_effect
+    mock_list_sources.side_effect = list_sources_side_effect
